+++ python/pydantable/__init__.py
@@ -18,4 +18,4 @@ from .expressions import Expr
 from .schema import Schema
 
 __all__ = ["DataFrame", "DataFrameModel", "Expr", "Schema", "pandas", "pyspark"]
-__version__ = "0.10.0"
+__version__ = "0.11.0"